# doc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notification Service API",
        "description": "Inbox notifications, notification templates, internal notification commands, and health endpoints.",
        "version": "1.0.0"
    },
    "paths": {
        "/notification/health": {
            "get": {
                "operationId": "get_notification_health",
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/notification/internal/request-access": {
            "post": {
                "operationId": "post_notification_internal_request_access",
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/notification/inbox": {
            "get": {
                "operationId": "get_notification_inbox",
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "post_notification_inbox_create",
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/notification/inbox/{id}/read": {
            "patch": {
                "operationId": "patch_notification_mark_read",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/notification/settings/template/request-access": {
            "get": {
                "operationId": "get_notification_request_access_template_get",
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "put": {
                "operationId": "put_notification_request_access_template_update",
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/notification/inbox/{id}": {
            "put": {
                "operationId": "put_notification_inbox_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "delete_notification_inbox_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}